requirements.txt
langgraph==0.1.3
openai==1.3.0
python-dotenv==1.0.0
rich==13.7.0
pydantic==2.5.0
langchain-openai==0.0.8
langchain-core==0.1.0
